A lot of things.
Created an authorization. Token replacement is not fully implemented yet. Created a blank for email, haven't decided if I will use it yet. Connected multer. Now with patch method user can add a link to the avatar. express distributes statics for it. I'm starting the final stage. Almost all the functionality that I had planned is close to completion. It remains only to add images for goods and possibly reset password for user. Improve methods where necessary, Write documentation for each module. And test the functionality of each endpoint, already started to test the user.